Fix progress bar, make public function names available to client
Earlier, the make_public function of the microservices was called each time the microservice was called, because the pointers to functions were not being stored in bus.microservices[microservice_handle]. Now, bus.microservices[microservice_handle] stores two dictionaries called 'public' and 'private', which store the function names as well as the pointers. This works because we have only one instance of each microservice running on the server at any given time, so the function pointers stored in these dictionaries will be bound to that instance.
Also, list_microservices called by the client returns a list of public functions along with the microservice identifiers.